Optimisation useCase1 (#57)
Optimisation of useCase1
Genotype.of() issue solved
Java version set to 10

+++ project/src/main/java/useCase/UseCase1.java
@@ -118,7 +118,7 @@ public class UseCase1 extends UseCase {
 
 		// then we add the new decision variables
 		for (int sat = 0; sat < nbSatMax; sat++) {
-			String newName = decVarAnomaly.getName() + Integer.toString(sat);
+			String newName = decVarAnomaly.getName() + Double.toString(sat);
 			DecisionVariable newDecVarAnomaly = new DecisionVariable(decVarAnomaly.getClass(), newName,
 					decVarAnomaly.getMin(), decVarAnomaly.getMax());
 			this.variablesList.add(newDecVarAnomaly);
